+++ sandbox/+tydex/+parsers/ExampleParser.m
@@ -1,0 +1,11 @@
+classdef ExampleParser < tydex.Parser
+    methods (Access = protected)
+        [measurements, bins, binvalues] = parse(obj,file)
+    end
+    methods
+        function [measurements, bins, binvalues] = run(obj, file)
+            [measurements, bins, binvalues] = parse(obj, file);
+            measurements = measurements.index();
+        end
+    end
+end