+++ src/+ui/TyreFitterSolverSettingsPanel.m
@@ -10,13 +10,14 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
         Panel                       matlab.ui.container.Panel
         Grid                        matlab.ui.container.GridLayout
         AlgorithmDropdown           matlab.ui.control.DropDown
-        MaxFunEvalEditField         matlab.ui.control.EditField
-        MaxIterEditField            matlab.ui.control.EditField
+        MaxFunEvalEditField         matlab.ui.control.NumericEditField
+        MaxIterEditField            matlab.ui.control.NumericEditField
         ParpoolButton               matlab.ui.control.StateButton
+        DownsampleFactorSpinner     matlab.ui.control.Spinner
     end
     
     properties (Dependent, Access = public)
-        OptimizerOptions optim.options.Fmincon
+        OptimizerOptions struct
     end
     
     properties (Access = private)
@@ -37,23 +38,29 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
             opts = obj.OptimizerOptions;
             s = obj.Settings.Fitter;
             try
-                opts.Algorithm = ...
-                    obj.AlgorithmDropdown.Value;
-                opts.MaxFunctionEvaluations = str2double(...
-                    obj.MaxFunEvalEditField.Value);
-                opts.MaxIterations =  str2double(...
-                    obj.MaxIterEditField.Value);
-                opts.UseParallel = logical(...
-                    obj.ParpoolButton.Value);
+                algorithm = obj.AlgorithmDropdown.Value;
+                opts.Algorithm = algorithm;
+                s.OptimizerSettings.Algorithm = opts.Algorithm;
+
+                maxFunEval = obj.MaxFunEvalEditField.Value;
+                opts.MaxFunEvals = maxFunEval;
+                s.OptimizerSettings.MaxFunEvals = maxFunEval;
+
+                maxIter = obj.MaxIterEditField.Value;
+                opts.MaxIter =  maxIter;
+                s.OptimizerSettings.MaxIterations = maxIter;
+
+                useParallel = logical(obj.ParpoolButton.Value);
+                opts.UseParallel = useParallel;
+                s.OptimizerSettings.UseParallel = useParallel;
+
+                downsampleFactor = obj.DownsampleFactorSpinner.Value;
+                s.DownsampleFactor = downsampleFactor;
             catch cause
                 exception = exceptions.InvalidSolverOptions();
                 exception = addCause(exception, cause);
                 throw(exception)
             end
-            s.OptimizerSettings.Algorithm = opts.Algorithm;
-            s.OptimizerSettings.MaxFunctionEvaluations = opts.MaxFunctionEvaluations;
-            s.OptimizerSettings.MaxIterations = opts.MaxIterations;
-            s.OptimizerSettings.UseParallel = opts.UseParallel;
             e = events.FitterSettingsChangedEventData(opts);
             obj.OptimizerOptions = opts;
             notify(obj, 'SettingsChanged', e)
@@ -65,6 +72,11 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
             obj.Position = [0 0 400 400];
             
             obj.Settings = settings.AppSettings();
+
+            if isempty(obj.OptimizerOptions)
+                opts = magicformula.v61.Fitter.initOptimizerOptions();
+                obj.OptimizerOptions = opts;
+            end
             
             g = uigridlayout(obj, ...
                 'RowHeight', {'1x'}, ...
@@ -74,7 +86,7 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
                 'Title', 'Optimization Settings', ...
                 'BorderType', 'none');
             obj.Grid = uigridlayout(obj.Panel, ...
-                'RowHeight', repmat({20}, 1, 4), ...
+                'RowHeight', repmat({20}, 1, 5), ...
                 'ColumnWidth', repmat({'fit'}, 1, 2), ...
                 'Padding', 5*ones(1,4), ...
                 'ColumnSpacing', 10);
@@ -90,11 +102,13 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
                 'ValueChangedFcn', @obj.onSettingsChanged);
             uilabel(obj.Grid, 'Text', 'Algorithm');
             
-            obj.MaxFunEvalEditField = uieditfield(obj.Grid, ...
+            obj.MaxFunEvalEditField = uieditfield(obj.Grid, 'numeric', ...
+                'Limits', [1 inf], ...
                 'ValueChangedFcn', @obj.onSettingsChanged);
             uilabel(obj.Grid, 'Text', 'MaxFunEvals');
             
-            obj.MaxIterEditField = uieditfield(obj.Grid, ...
+            obj.MaxIterEditField = uieditfield(obj.Grid, 'numeric', ...
+                'Limits', [1 inf], ...
                 'ValueChangedFcn', @obj.onSettingsChanged);
             uilabel(obj.Grid, 'Text', 'MaxIter');
             
@@ -102,13 +116,23 @@ classdef TyreFitterSolverSettingsPanel < matlab.ui.componentcontainer.ComponentC
                 'Text', 'Enable', ...
                 'ValueChangedFcn', @obj.onSettingsChanged);
             uilabel(obj.Grid, 'Text', 'UseParallel');
+
+            tooltip = 'Increase downsampling to reduce fit time.';
+            obj.DownsampleFactorSpinner = uispinner(obj.Grid, ...
+                'Limits', [1 1E5], ...
+                'Tooltip', tooltip, ...
+                'ValueChangedFcn', @obj.onSettingsChanged);
+            uilabel(obj.Grid, 'Text', 'Downsampling', ...
+                'Tooltip', tooltip);
         end
         function update(obj)
+            s = obj.Settings.Fitter;
             opts = obj.OptimizerOptions;
             obj.AlgorithmDropdown.Value = opts.Algorithm;
-            obj.MaxFunEvalEditField.Value = num2str(opts.MaxFunEvals);
-            obj.MaxIterEditField.Value = num2str(opts.MaxIter);
+            obj.MaxFunEvalEditField.Value = opts.MaxFunEvals;
+            obj.MaxIterEditField.Value = opts.MaxIter;
             obj.ParpoolButton.Value = opts.UseParallel;
+            obj.DownsampleFactorSpinner.Value = s.DownsampleFactor;
         end
     end
 end